backend/requirements.txt
asgiref==3.8.1
<<<<<<< HEAD:phisher_backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
joblib==1.4.2
numpy==2.2.2
scikit-learn==1.6.1
scipy==1.15.1
sqlparse==0.5.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
=======
Django==5.1.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
sqlparse==0.5.3
>>>>>>> 65abdb8ca5d222bec64873dc7dc58f245f72e172:backend/requirements.txt
